requirements.txt
requests==2.32.3
python-slugify==8.0.4
